Large update to wounds. Other small updates.
The four difficulties are Minor, Significant, ??, Legendary. Significant is the only one used here.

Wounds are generalized to be physical or mental, and explained much more fully. Short and long rests added.

Slight rewording on the topic of descriptions.

Horizontal breaks between subsections. I suspect that subheadings are not necessary. Maybe a doodle or something in these places in the finished text?