Updated terminology in rolls explanations.
Updated the difficulty scale to Minor, Significant, Incredible, Legendary.

Changed "difficulty" to "level" to line up with monsters.

d6-d12 -> d4-d10.

Noted that default rolls are d6, and that legendary rolls are the rarest. 

Got rid of a Ranger (!) reference in the sewer example.

COnfrontations are not yet updated.

It was previously the case that spells could demand rolls other than the bloodline they came from -- that's no longer the case.