Made room for new spells!
The existing text seems fine for now. It's possible that we want advancement before a session, so that you get one right away? After a session makes it easier to know what your character just did, though, and what it wants. 

We need some spells, obviously. I'm working on it.