Up to date.
A few sample spells added for each people. These are not great spells, and not particularly orthogonal, but they are at least something playable in time for the playtest. We can come up with more, better spells later. 

I did try to come up with a variety of styles of spells so we can see what works. Some are static. SOme you roll for. Some explicitly modify rolls (analogous to stunts -- that's likely an analogy worth exploring).