Up to date.
OK, not quite, but it's good enough for now. I need to come up with an incredible Goblin roll that you would make against a dragon. Ideally, neither sneaking nor shooting. Maybe the dragon keeps something important behind a locked door, but it's too cocky to have a good lock?