Cut descriptions influencing roll level
I like having descriptions be a place for players to write what they feel is important about their PC, even if that means it starts off pretty sparse early in the campaign. The alternative seems to be encouraging players to minmax their descriptions, filling them with expertises to gain bonuses on rolls.