Sketched out confrontations
This essay is sketched out paragraph by paragraph, more or less. I should be able to make it pretty tomorrow. I think after this we just need to clean up some miscellaneous odds and ends, come up with a few more spells, and get Alex to write the GM chapter.