worked on spells some
Added in a few spells, and notes on what the spells header should say. I'm not sure how much text I want for each spell, but I think I do want a few example uses to get imaginations flowing. How much mechanical clarification and restriction is necessary is probably something that'll need to be tested, as will be how explicit we need to be with when the GM can call for a roll during a spell's use.