paragraph on optional weaknesses
Wrote a bit about how you can, if you so choose, play a PC who is persuasive but doesn't lie, or who is sneaky but doesn't carry a bow. 

Also, noted that it's awkward that sample PCs are at the very end. Shouldn't we introduce them early so we can use them to explain things?