Explained where spells come from
PCs level up at the START of each session, not the end. This lets you choose a bonus and then use it immediately, which is probably fun. It also means that you can take a spell right away if you want. 

Next up: for the sample PCs, include the first level up.