wrote humans intro
I swapped the order of the third and fourth paragraphs in each of the essays. The human essay flows better that way, and the rest seem pretty agnostic. 

The human essay is the longest by a fair margin. The others should perhaps be fleshed out a bit. Maybe name a city or two for each people? 

The human capital city does not yet have a name.